remove event queue
